Document the model using README
I created instructions on how to run the model in the README markdown file. I also deleted the test run section in the model.py module. It seems like, with the way that the model is structured right now, that it is impractical to run the model for time steps of at least 30.

Next step: analyzing model runtime.