Analyzed model run time and created another model
I found that the time it takes for the original model to (1) create particles in a particular time step and, in the same time step, (2) perform calculations, both grow exponentially with the number of time steps. This creates a prohibitively computationally expensive model, so I developed a different model in which the number of particles are assumed to grow linearly with each time step when the total number of particles are greater than 1000.